requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.7
django-allauth==0.50.0
django-cors-headers==3.13.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.13.1
drf-yasg==1.20.0
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.0
packaging==21.3
Pillow==9.1.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.8
python-dateutil==1.5
python-decouple==3.5
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.0.0
